requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pandas>=1.4.0
PyYAML>=6.0
scikit-learn>=1.0.0
seaborn>=0.11.0
ultralytics>=8.0.0
tqdm>=4.64.0
torch>=1.13.0
torchvision>=0.14.0 
gdown>=4.5.0